+++ hcga/features/basal_nodes.py
@@ -6,7 +6,10 @@ from ..feature_class import FeatureClass, InterpretabilityScore
 
 featureclass_name = "BasalNodes"
 
-"Basal nodes are nodes which have in degree equal to zero"
+"""
+Basal nodes are nodes which have in degree equal to zero. Attracting nodes are 
+nodes which have out degree equal to zero
+"""
 
 class BasalNodes(FeatureClass):
     """Basal nodes class."""
@@ -56,6 +59,47 @@ class BasalNodes(FeatureClass):
             "exp_basal_edge",
             lambda graph: exp_basal_edge(graph),
             "The distribution of the expected number of in-edges of each node with basal nodes",
+            InterpretabilityScore(3),
+            statistics = "centrality",
+        )
+        
+        def attracting_nodes_func(graph):
+            out_degrees = dict(graph.out_degree)
+            return [i for i in out_degrees if out_degrees[i] == 0]
+        
+        self.add_feature(
+            "n_attracting_nodes",
+            lambda graph: len(attracting_nodes_func(graph)),
+            "The number of basal nodes",
+            InterpretabilityScore(3),
+        )
+            
+        self.add_feature(
+            "attracting_degrees",
+            lambda graph: [dict(graph.in_degree)[i] for i in attracting_nodes_func(graph)],
+            "The distribution of degrees of attracting nodes",
+            InterpretabilityScore(3),
+            statistics = "centrality",
+        )
+        
+        n_attracting_edges = lambda graph: sum([dict(graph.in_degree)[i] for i in attracting_nodes_func(graph)])
+        
+        self.add_feature(
+            "n_attracting_edges",
+            n_attracting_edges,
+            "The total number of edges connected to attracting nodes",
+            InterpretabilityScore(3),
+        )
+        
+        def exp_attracting_edge(graph):
+            out_degs = list(dict(graph.out_degree).values())
+            r = sum([dict(graph.in_degree)[i] for i in attracting_nodes_func(graph)])/(graph.number_of_edges())
+            return [i*r for i in out_degs]
+        
+        self.add_feature(
+            "exp_attracting_edge",
+            lambda graph: exp_attracting_edge(graph),
+            "The distribution of the expected number of out-edges of each node with attracting nodes",
             InterpretabilityScore(3),
             statistics = "centrality",
         )
